% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-export.R
\name{create_zipfile}
\alias{create_zipfile}
\title{Create Zip File with MS ACCESS Database And Exported CSV files}
\usage{
create_zipfile(db_path, target_dir = dirname(db_path), dbg = TRUE)
}
\arguments{
\item{db_path}{path to MS Access DB (default: default_access_db())}

\item{target_dir}{directory. By default a new zip file is created in the same 
directory as mdb resides in. The new directory has the same name as the 
database file but with file extension ".zip" (default: dirname(db_path))}

\item{dbg}{debug information (default: TRUE)}
}
\description{
Create Zip File with MS ACCESS Database And Exported CSV files
}
