% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-export.R
\name{default_access_db}
\alias{default_access_db}
\title{Default Path to MS ACCESS Database}
\usage{
default_access_db()
}
\value{
path to MS ACCESS DB
}
\description{
Default Path to MS ACCESS Database
}
\examples{
default_access_db()
}
