% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-export.R
\name{default_csv_dir}
\alias{default_csv_dir}
\title{Helper Function: Default Csv Directory}
\usage{
default_csv_dir(db_path)
}
\arguments{
\item{db_path}{path to MS Access DB (default: deufault_access_db())}
}
\value{
csv directory but with all "." replaced with
with "_" (e.g. "qmra-db_accdb" instead of "qmra-db.accdb")
}
\description{
Helper Function: Default Csv Directory
}
\examples{
db_path <- default_access_db()
default_csv_dir(db_path)
}
