% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-export.R
\name{default_target_dir}
\alias{default_target_dir}
\title{Helper Function: Default Target Directory}
\usage{
default_target_dir(db_path, csv_dir = default_csv_dir(db_path))
}
\arguments{
\item{db_path}{path to MS Access DB (default: deufault_access_db())}

\item{csv_dir}{csv directory (default: default_csv_dir())}
}
\value{
subdirectory in MS ACCESS DB directory but with all "." replaced with
with "_" (e.g. "qmra-db_accdb" instead of "qmra-db.accdb")
}
\description{
Helper Function: Default Target Directory
}
\examples{
db_path <- default_access_db()
default_target_dir(db_path)
}
