% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-export.R
\name{dump_access_db}
\alias{dump_access_db}
\title{Dump MS ACCESS Database To CSV}
\usage{
dump_access_db(db_path = default_access_db())
}
\arguments{
\item{db_path}{path to MS Access DB (default: deufault_access_db())}
}
\description{
for exporting tables and queries from MS ACCESS DB into 
.csv files (one for each table/query) into subfolder "qmra-db_accdb"
}
