% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_db_content.R
\name{import_database_content}
\alias{import_database_content}
\title{Import Database Content From CSV Files}
\usage{
import_database_content(csv_dir = system.file("database/qmra-db_accdb",
  package = "qmra.db"), encoding = "WINDOWS-1252")
}
\arguments{
\item{csv_dir}{path to csv directory with database output (default: 
system.file("database/qmra-db_accdb", package = "qmra.db"))}

\item{encoding}{encoding of csv files (default: "WINDOWS-1252")}
}
\value{
list with multiple data.frames with database content
}
\description{
Import Database Content From CSV Files
}
\examples{
\dontrun{
db_content <- import_database_content()
str(db_content, 1)
}
}
